rooksafe/requirements.txt
amqp==5.3.1
asgiref==3.8.1
beautifulsoup4==4.12.3
billiard==4.2.1

certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Django==5.1.3
google-api-python-client
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Faker==33.0.0
finnhub-python==2.4.22
fonttools==4.55.0
frozendict==2.4.6
gevent==24.11.1
greenlet==3.1.1
gunicorn==23.0.0
html5lib==1.1
idna==3.10
kiwisolver==1.4.7
kombu==5.4.2
lxml==5.3.0
matplotlib==3.9.3
msgpack==1.1.0
multitasking==0.0.11
numpy==2.1.3
packaging==24.2
pandas==2.2.3
peewee==3.17.8
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.0
requests==2.32.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
#uwsgi      #ARIEL NO ME MATES PERO COMO AUN NO ESTAMOS EN PRODUCCION, ME TOME EL ATREVIMIENTO
vine==5.1.0
waitress==3.0.2
wcwidth==0.2.13
webencodings==0.5.1
websocket==0.2.1
websocket-client==1.8.0
yfinance==0.2.50
zope.event==5.0
zope.interface==7.2
feedparser
langdetect
daphne